requirements.txt
absl-py==0.6.1
appnope==0.1.0
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
conda==4.3.16
cycler==0.10.0
dask==0.20.2
decorator==4.3.0
defusedxml==0.5.0
dnspython==1.15.0
entrypoints==0.2.3
eventlet==0.24.1
ffmpeg==1.4
Flask==1.0.2
Flask-SocketIO==3.0.2
gast==0.2.0
greenlet==0.4.15
grpcio==1.16.1
h5py==2.8.0
idna==2.7
imageio==2.4.1
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
monotonic==1.5
moviepy==0.2.3.5
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.2
numpy==1.15.4
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pycosat==0.6.3
Pygments==2.2.0
pyparsing==2.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.13
python-dateutil==2.7.5
python-engineio==2.3.2
python-socketio==2.0.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.20.1
ruamel.yaml==0.15.78
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
